% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comparison_dataset}
\alias{comparison_dataset}
\title{Simulated normal dataset with mean=5 and variance=1}
\format{
A vector
}
\source{
simulations
}
\usage{
comparison_dataset
}
\description{
A dataset that contains simulated data to reproduce a figure in our manuscript
}
\keyword{datasets}
