% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expression_data}
\alias{expression_data}
\title{Log expression data for select genes}
\format{
A data frame with 1200 rows and 3 variables:
\describe{
\item{gene}{gene identifier (not meaningful)}
\item{specimen}{autism or control}
\item{log_count}{the logged gene expression count}
...
}
}
\source{
\url{https://www.ebi.ac.uk/gxa/experiments/E-GEOD-30573/Results}
}
\usage{
expression_data
}
\description{
A dataset that contains log expression data for randomly selected genes for two
patients, one with autism and one control.
}
\keyword{datasets}
