% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_qqboxplot.R
\name{geom_qqboxplot}
\alias{geom_qqboxplot}
\title{A modification of the boxplot with information about the tails}
\usage{
geom_qqboxplot(
  mapping = NULL,
  data = NULL,
  stat = "qqboxplot",
  position = "dodge2",
  ...,
  outlier.colour = NULL,
  outlier.color = NULL,
  outlier.fill = NULL,
  outlier.shape = 19,
  outlier.size = 1.5,
  outlier.stroke = 0.5,
  outlier.alpha = NULL,
  notch = FALSE,
  notchwidth = 0.5,
  varwidth = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
A modification of the boxplot with information about the tails
}
