% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicators}
\alias{indicators}
\title{World Bank indicator data for Labor Force participation rates}
\format{
A data frame with 612 rows and 7 variables:
\describe{
\item{Country Name}{name of country}
\item{Country Code}{unique country identifier (string)}
\item{Series Name}{Specifies male/female}
\item{Series Code}{unique identifier for series}
\item{year}{year for data}
\item{indicator}{participation rate in percents}
\item{log_indicator}{the log of the participation rate}
...
}

A data frame with 12800 rows and 5 variables:
\describe{
\item{orientation}{1 to 8, specifies the orientation of the grating}
\item{nspikes}{number of spikes for a single trial of 1.28 seconds for a particular orientation}
\item{region}{region of the brain where the neuron is located}
...
}

A data frame with 13731 rows and 3 variables:
\describe{
\item{ecephys_structure_acronym}{acronym for population location}
\item{fr_type}{situation under which firing rate was recorded}
\item{rate}{the firing rate}
...
}
}
\source{
\url{https://datacatalog.worldbank.org/dataset/world-development-indicators}

\url{https://CRCNS.org}

\url{https://allensdk.readthedocs.io/en/latest/visual_coding_neuropixels.html}
}
\usage{
indicators

indicators

indicators
}
\description{
A dataset that contains participation rates (\%) for ages 15-24, separated
by gender, and measured in the years 2008, 2012, and 2017

A dataset that contains the number of spikes for neurons across several
possible orientations of a grating

A dataset that contains populations of neurons from CA1 and LM and their
firing rates for three situations: base firing rate, dot motion, and
drifting gradient.  Each row represents a neuron
}
\keyword{datasets}
