% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_brain_data}
\alias{population_brain_data}
\title{Neuron population firing data}
\format{
A data frame with 13731 rows and 3 variables:
\describe{
\item{ecephys_structure_acronym}{acronym for population location}
\item{fr_type}{situation under which firing rate was recorded}
\item{rate}{the firing rate}
...
}
}
\source{
\url{https://allensdk.readthedocs.io/en/latest/visual_coding_neuropixels.html}
}
\usage{
population_brain_data
}
\description{
A dataset that contains populations of neurons from CA1 and LM and their
firing rates for three situations: base firing rate, dot motion, and
drifting gradient.  Each row represents a neuron
}
\keyword{datasets}
