% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{Simulated t-distributions to show use of q-q boxplots}
\format{
A data frame with 4500 rows and 2 variables:
\describe{
\item{y}{a value simulated from a distribution}
\item{group}{a string specifying the distribution from which the y value
is drawn}
...
}
}
\source{
simulations
}
\usage{
simulated_data
}
\description{
A dataset that contains simulated data to reproduce the simulated data
figures used in our manuscript
}
\keyword{datasets}
