% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spike_data}
\alias{spike_data}
\title{Neuron spiking data for neural tuning orientation}
\format{
A data frame with 12800 rows and 5 variables:
\describe{
\item{orientation}{1 to 8, specifies the orientation of the grating}
\item{nspikes}{number of spikes for a single trial of 1.28 seconds for a particular orientation}
\item{region}{region of the brain where the neuron is located}
...
}
}
\source{
\url{https://CRCNS.org}
}
\usage{
spike_data
}
\description{
A dataset that contains the number of spikes for neurons across several
possible orientations of a grating
}
\keyword{datasets}
