% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_qqboxplot.R, R/stat_qqboxplot.R
\name{geom_qqboxplot}
\alias{geom_qqboxplot}
\alias{stat_qqboxplot}
\title{A modification of the boxplot with information about the tails}
\usage{
geom_qqboxplot(
  mapping = NULL,
  data = NULL,
  stat = "qqboxplot",
  position = "dodge2",
  ...,
  outlier.colour = NULL,
  outlier.color = NULL,
  outlier.fill = NULL,
  outlier.shape = 19,
  outlier.size = 1.5,
  outlier.stroke = 0.5,
  outlier.alpha = NULL,
  notch = FALSE,
  notchwidth = 0.5,
  varwidth = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_qqboxplot(
  mapping = NULL,
  data = NULL,
  geom = "qqboxplot",
  position = "dodge2",
  ...,
  coef = 1.5,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  reference_dist = "norm",
  confidence_level = 0.95,
  compdata = NULL
)
}
\arguments{
\item{reference_dist}{Specifies theoretical reference distribution.}

\item{confidence_level}{Sets confidence level for deviation whisker
confidence bands}

\item{compdata}{specifies a data set to use as the reference distribution.
If compdata is not NULL, the argument reference_dist will be ignored.}
}
\description{
A modification of the boxplot with information about the tails
}
\section{Description}{

The qq-boxplot inherits its summary statistics from the boxplot.  See
\code{\link[=geom_boxplot]{geom_boxplot()}} for details.  The qq-boxplot differs from the boxplot
by using more informative whiskers than the regular boxplot.

The vertical position of the whiskers can be interpreted as it is in the
boxplot, and the maximal vertical value is chosen as it is done in the
regular boxplot.  The horizontal positioning of the whiskers indicates the
deviation of the data set of interest from some reference data set
(specified as either a theoretical distribution or an actual data set).
Taking the central vertical axis of the boxplot as being zero, deviations
to the right indicate that those values are larger than the
corresponding data points in the reference data set, where two data points
correspond if their quantiles match.  Deviations to the left indicate that
the values are smaller than their corresponding data points.  Consider a
situation where your data set has fatter tails than the normal distribution.
When the reference distribution is the normal distribution, then the whiskers
below the box will be left of the central axis (the left tail values are
smaller than they ought to be) and the whiskers above the box will be right
of the central axis (the right tail values are larger than the ought to be).

In order to compare the data set of interest to the reference data set, they
must be on the same scale.  The qq-boxplot uses Tukey's g-h distribution
to determine the appropriate scaling factor.

Much of the code here is a modification of the geom_boxplot() code.
}

\section{Computed variables}{

\code{stat_qqboxplot()} provides the following variables, some of which depend on the orientation:
\describe{
\item{width}{width of boxplot}
\item{ymin \emph{or} xmin}{lower whisker = smallest observation greater than or equal to lower hinge - 1.5 * IQR}
\item{lower \emph{or} xlower}{lower hinge, 25\% quantile}
\item{notchlower}{lower edge of notch = median - 1.58 * IQR / sqrt(n)}
\item{middle \emph{or} xmiddle}{median, 50\% quantile}
\item{notchupper}{upper edge of notch = median + 1.58 * IQR / sqrt(n)}
\item{upper \emph{or} xupper}{upper hinge, 75\% quantile}
\item{ymax \emph{or} xmax}{upper whisker = largest observation less than or equal to upper hinge + 1.5 * IQR}
}
}

