% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicators}
\alias{indicators}
\title{World Bank indicator data for Labor Force participation rates}
\format{
A data frame with 612 rows and 7 variables:
\describe{
\item{Country Name}{name of country}
\item{Country Code}{unique country identifier (string)}
\item{Series Name}{Specifies male/female}
\item{Series Code}{unique identifier for series}
\item{year}{year for data}
\item{indicator}{participation rate in percents}
\item{log_indicator}{the log of the participation rate}
...
}
}
\source{
\url{https://www.worldbank.org/en/home}
}
\usage{
indicators
}
\description{
A dataset that contains participation rates (\%) for ages 15-24, separated
by gender, and measured in the years 2008, 2012, and 2017
}
\keyword{datasets}
